Do not run suspense.tsx with mocha
