+++ compat/src/suspense.d.ts
@@ -1,0 +1,13 @@
+import { Component } from '../../src';
+
+//
+// Suspense/lazy
+// -----------------------------------
+export function lazy<T>(loader: () => Promise<{default: T}>): T;
+
+interface SuspenseProps {
+  children?: preact.ComponentChildren;
+  fallback: preact.ComponentChildren;
+}
+
+export abstract class Suspense extends Component<SuspenseProps> {}
